+++ sky-server/src/main/java/com/sky/service/CategoryService.java
@@ -54,5 +54,5 @@ public interface CategoryService {
      * 根据id删除分类
      * @param id
      */
-    void delete(Integer id);
+    void delete(Long id);
 }
